// client/src/app/profaktura/korisnik/korisnik.model.ts
export interface Korisnik {
  br_pasos: number;
  br_telefona: number;
  broj: number;
  datum_rodj: string;
  id_adrese: number;
  id_drzave: number;
  id_grada: number;
  ime_prezime: string;
  jmbg: string;
  naziv_drzave: string;
  naziv_grada: string;
  sifra_programa: number;
  ulica: string;
}